p4/requirements.txt
matplotlib
scipy
scikit-image
scikit-learn
tqdm
numpy
pillow